refactor: move VFSSemaphore from core/ to lib/ — kernel-authored standard
VFSSemaphore is consumed by factory and backends, not by kernel directly.
Move implementation to lib/ (tier-neutral) and extract protocol to
contracts/protocols/semaphore.py.

- Create contracts/protocols/semaphore.py (VFSSemaphoreProtocol)
- Create lib/semaphore.py (PythonVFSSemaphore, RustVFSSemaphore, factory)
- Delete core/semaphore.py
- Update 2 production importers (factory/_bricks.py, cas_local.py)
- Update 4 test importers
- Move test_vfs_semaphore.py from tests/unit/core/ to tests/unit/lib/
- Update lock-architecture.md package references
