+++ src/nexus/lib/semaphore.py
@@ -26,40 +26,11 @@ import logging
 import threading
 import time
 import uuid
-from typing import Any, Protocol, cast, runtime_checkable
+from typing import Any, cast
+
+from nexus.contracts.protocols.semaphore import VFSSemaphoreProtocol
 
 logger = logging.getLogger(__name__)
-
-# ---------------------------------------------------------------------------
-# Protocol
-# ---------------------------------------------------------------------------
-
-
-@runtime_checkable
-class VFSSemaphoreProtocol(Protocol):
-    """Structural interface shared by Rust and Python implementations."""
-
-    def acquire(
-        self,
-        name: str,
-        max_holders: int,
-        timeout_ms: int = 0,
-        ttl_ms: int = 30_000,
-    ) -> str | None: ...
-
-    def release(self, name: str, holder_id: str) -> bool: ...
-
-    def extend(self, name: str, holder_id: str, ttl_ms: int = 30_000) -> bool: ...
-
-    def info(self, name: str) -> dict | None: ...
-
-    def force_release(self, name: str) -> bool: ...
-
-    def stats(self) -> dict: ...
-
-    @property
-    def active_semaphores(self) -> int: ...
-
 
 # ---------------------------------------------------------------------------
 # Internal data classes
